+++ fitConnect/src/main/java/com/example/fitconnect/controller/chat/ChatController.java
@@ -1,9 +1,7 @@
 package com.example.fitconnect.controller.chat;
 
 
-import com.example.fitconnect.config.annotation.CurrentUserId;
-import com.example.fitconnect.domain.chat.domain.ChatMessage;
-import com.example.fitconnect.domain.chat.dto.ChatMessageRegistrationDto;
+import com.example.fitconnect.dto.chat.request.ChatMessageRegistrationDto;
 import com.example.fitconnect.dto.chatMessage.response.ChatMessageResponseDto;
 import com.example.fitconnect.service.chat.chatMessage.ChatMessageCreationService;
 import lombok.RequiredArgsConstructor;
@@ -11,11 +9,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestBody;
 
 @Controller
 @RequiredArgsConstructor
