+++ fitConnect/src/main/java/com/example/fitconnect/controller/chat/ChatMessageController.java
@@ -2,20 +2,14 @@ package com.example.fitconnect.controller.chat;
 
 
 import com.example.fitconnect.config.annotation.CurrentUserId;
-import com.example.fitconnect.config.service.CommonService;
-import com.example.fitconnect.domain.chat.domain.ChatMessage;
-import com.example.fitconnect.domain.chat.dto.ChatMessageUpdateDto;
+import com.example.fitconnect.dto.chat.request.ChatMessageUpdateDto;
 import com.example.fitconnect.dto.chatMessage.response.ChatMessageResponseDto;
 import com.example.fitconnect.service.chat.chatMessage.ChatMessageDeleteService;
 import com.example.fitconnect.service.chat.chatMessage.ChatMessageFindService;
 import com.example.fitconnect.service.chat.chatMessage.ChatMessageUpdateService;
-import com.nimbusds.oauth2.sdk.Message;
-import jakarta.servlet.http.HttpSession;
 import java.util.List;
-import java.util.Map;
 import jdk.jshell.Snippet.Status;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.repository.query.Param;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
