+++ fitConnect/src/main/java/com/example/fitconnect/dto/event/request/RecruitmentPolicyDto.java
@@ -1,4 +1,4 @@
-package com.example.fitconnect.domain.event.dto;
+package com.example.fitconnect.dto.event.request;
 
 import com.example.fitconnect.domain.event.domain.RecruitmentPolicy;
 import java.time.LocalDateTime;
