+++ fitConnect/src/main/java/com/example/fitconnect/service/chat/chatMessage/ChatMessageUpdateService.java
@@ -4,7 +4,7 @@ import static com.example.fitconnect.config.error.ErrorMessages.*;
 
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.chat.domain.ChatMessage;
-import com.example.fitconnect.domain.chat.dto.ChatMessageUpdateDto;
+import com.example.fitconnect.dto.chat.request.ChatMessageUpdateDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
