+++ fitConnect/src/main/java/com/example/fitconnect/service/chat/chatRoom/ChatRoomCreationService.java
@@ -5,7 +5,7 @@ import com.example.fitconnect.config.error.ErrorMessages;
 import com.example.fitconnect.config.exception.BusinessException;
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.chat.domain.ChatRoom;
-import com.example.fitconnect.domain.chat.dto.ChatRoomRegistrationDto;
+import com.example.fitconnect.dto.chat.request.ChatRoomRegistrationDto;
 import com.example.fitconnect.domain.event.domain.ExerciseEvent;
 import com.example.fitconnect.domain.user.domain.User;
