+++ fitConnect/src/main/java/com/example/fitconnect/service/chat/chatRoom/ChatRoomUpdateService.java
@@ -3,7 +3,7 @@ package com.example.fitconnect.service.chat.chatRoom;
 import com.example.fitconnect.config.error.ErrorMessages;
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.chat.domain.ChatRoom;
-import com.example.fitconnect.domain.chat.dto.ChatRoomUpdateDto;
+import com.example.fitconnect.dto.chat.request.ChatRoomUpdateDto;
 import com.example.fitconnect.repository.chat.chatRoom.ChatRoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
