+++ fitConnect/src/main/java/com/example/fitconnect/service/event/ExerciseEventUpdateService.java
@@ -3,7 +3,7 @@ package com.example.fitconnect.service.event;
 import com.example.fitconnect.config.error.ErrorMessages;
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.event.domain.ExerciseEvent;
-import com.example.fitconnect.domain.event.dto.ExerciseEventUpdateDto;
+import com.example.fitconnect.dto.event.request.ExerciseEventUpdateDto;
 import com.example.fitconnect.repository.event.ExerciseEventRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
