+++ fitConnect/src/main/java/com/example/fitconnect/service/review/ReviewCreationService.java
@@ -4,12 +4,10 @@ import com.example.fitconnect.config.error.ErrorMessages;
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.event.domain.ExerciseEvent;
 import com.example.fitconnect.domain.review.Review;
-import com.example.fitconnect.domain.review.dto.ReviewRegistrationDto;
+import com.example.fitconnect.dto.review.request.ReviewRegistrationDto;
 import com.example.fitconnect.domain.user.domain.User;
 import com.example.fitconnect.dto.review.response.ReviewResponseDto;
-import com.example.fitconnect.repository.event.ExerciseEventRepository;
 import com.example.fitconnect.repository.review.ReviewRepository;
-import com.example.fitconnect.repository.user.UserRepository;
 import com.example.fitconnect.service.event.ExerciseEventFindService;
 import com.example.fitconnect.service.user.UserFindService;
 import jakarta.transaction.Transactional;
