+++ fitConnect/src/main/java/com/example/fitconnect/service/review/ReviewUpdateService.java
@@ -3,7 +3,7 @@ package com.example.fitconnect.service.review;
 import com.example.fitconnect.config.error.ErrorMessages;
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.review.Review;
-import com.example.fitconnect.domain.review.dto.ReviewUpdateDto;
+import com.example.fitconnect.dto.review.request.ReviewUpdateDto;
 import com.example.fitconnect.domain.user.domain.User;
 import com.example.fitconnect.repository.review.ReviewRepository;
 import com.example.fitconnect.repository.user.UserRepository;
