+++ fitConnect/src/main/java/com/example/fitconnect/service/user/UserUpdateService.java
@@ -3,7 +3,7 @@ package com.example.fitconnect.service.user;
 import com.example.fitconnect.config.error.ErrorMessages;
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.user.domain.User;
-import com.example.fitconnect.domain.user.dto.UserUpdateDto;
+import com.example.fitconnect.dto.user.request.UserUpdateDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
