+++ fitConnect/src/test/java/com/example/fitconnect/controller/chat/ChatMessageControllerTest.java
@@ -1,10 +1,7 @@
 package com.example.fitconnect.controller.chat;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import com.example.fitconnect.auth.service.JwtService;
-import com.example.fitconnect.config.service.CommonService;
-import com.example.fitconnect.domain.chat.dto.ChatMessageUpdateDto;
+import com.example.fitconnect.dto.chat.request.ChatMessageUpdateDto;
 import com.example.fitconnect.service.chat.chatMessage.ChatMessageDeleteService;
 import com.example.fitconnect.service.chat.chatMessage.ChatMessageFindService;
 import com.example.fitconnect.service.chat.chatMessage.ChatMessageUpdateService;
@@ -17,15 +14,12 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.MockMvcBuilder;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
