+++ fitConnect/src/test/java/com/example/fitconnect/service/review/ReviewCreationServiceTest.java
@@ -4,12 +4,12 @@ import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.event.domain.Category;
 import com.example.fitconnect.domain.event.domain.City;
 import com.example.fitconnect.domain.event.domain.ExerciseEvent;
-import com.example.fitconnect.domain.event.dto.EventDetailDto;
-import com.example.fitconnect.domain.event.dto.ExerciseEventRegistrationDto;
-import com.example.fitconnect.domain.event.dto.LocationDto;
-import com.example.fitconnect.domain.event.dto.RecruitmentPolicyDto;
+import com.example.fitconnect.dto.event.request.EventDetailDto;
+import com.example.fitconnect.dto.event.request.ExerciseEventRegistrationDto;
+import com.example.fitconnect.dto.event.request.LocationDto;
+import com.example.fitconnect.dto.event.request.RecruitmentPolicyDto;
 import com.example.fitconnect.domain.review.Review;
-import com.example.fitconnect.domain.review.dto.ReviewRegistrationDto;
+import com.example.fitconnect.dto.review.request.ReviewRegistrationDto;
 import com.example.fitconnect.domain.user.domain.User;
 import com.example.fitconnect.dto.review.response.ReviewResponseDto;
 import com.example.fitconnect.repository.event.ExerciseEventRepository;
