+++ fitConnect/src/test/java/com/example/fitconnect/service/user/UserRegisterServiceTest.java
@@ -4,9 +4,8 @@ import com.example.fitconnect.config.exception.BusinessException;
 import com.example.fitconnect.domain.user.domain.Role;
 import com.example.fitconnect.domain.user.domain.User;
 import com.example.fitconnect.domain.user.domain.UserBaseInfo;
-import com.example.fitconnect.domain.user.dto.UserRegistrationDto;
+import com.example.fitconnect.dto.user.request.UserRegistrationDto;
 import com.example.fitconnect.repository.user.UserRepository;
-import com.example.fitconnect.service.user.UserRegisterService;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.api.extension.ExtendWith;
